Working on updating Vue test utils to beta 29.
For now, we'll put off adding {sync: false}
explicitly to the tests because our tests pass without it and
vue test utils is planning on removing the option of {sync: true}. This
commit does add some await $nextTick() calls to the validated form and
validated input tests that make them pass when {sync: false} is explicitly
used (although we don't explicitly include {sync: false} at this time).

Also refactored the way validated input and validated form store and emit
their "is_valid" state. Instead of relying on watchers (which don't seem
to always work in tests after switching to beta 29), "is_valid" is now
explicitly stored in a reactive data member, and an event is emitted
whenever this state changes. For validated form, this involves attaching
event listeners to all the registered validated inputs inside it.