Add checkboxes and event for file selection
Begin work for batched operations by allowing users to select
SingleInstructorFiles. The child component emits a selected event now
to notify the parent when a user has marked the child for a batch
operation.

The list of selected files is stored in the parent component,
InstructorFiles, who listens for the select event.

Co-authored-by: Cayetano Wagner <cowagner@umich.edu>